fix: Correctly use t function
