+++ packages/cozy-harvest-lib/src/services/softDeleteOrRestoreAccounts.js
@@ -5,22 +5,11 @@ import {
   fetchAccountsForCipherId,
   fetchKonnectorFromAccount,
   fetchTriggersFromAccount,
-  updateAccountsAuth
+  updateAccountsAuth,
+  getT
 } from './utils'
 import get from 'lodash/get'
 import { ensureTrigger } from '../connections/triggers'
-const Polyglot = require('node-polyglot')
-
-const supportedLocales = ['en', 'fr', 'es']
-
-const getT = () => {
-  const locale =
-    supportedLocales.find(l => l === process.env.COZY_LOCALE) || 'en'
-  const locales = require(`../locales/${locale}.json`)
-  const polyglot = new Polyglot()
-  polyglot.extend(locales)
-  return polyglot.t.bind(polyglot)
-}
 
 const makeDecrypt = (vaultClient, orgKey) => encryptedVal =>
   decryptString(encryptedVal, vaultClient, orgKey)
